+++ sophie/syntax.py
@@ -46,12 +46,15 @@ class SimpleType(Expr):
 	def dispatch_token(self): raise NotImplementedError(type(self))
 
 class ValExpr(Expr):
-	is_volatile: bool
 	pass
 
 class PlainReference(Reference):
 	def head(self) -> slice: return self.nom.head()
-	def __str__(self): return "<ref:%s>"%self.nom.text
+	def __repr__(self): return "<ref:%s>"%self.nom.text
+
+class MemberReference(Reference):
+	def head(self) -> slice: return self.nom.head()
+	def __repr__(self): return "<my %s>"%self.nom.text
 
 class QualifiedReference(Reference):
 	space: Nom
@@ -96,6 +99,8 @@ class TypeCall(SimpleType):
 		symbol = self.ref.dfn
 		assert isinstance(symbol, TypeDeclaration)
 		return symbol.as_token()
+	def __repr__(self):
+		return "%s[%s]"%(self.ref, self.arguments) if self.arguments else repr(self.ref)
 
 class ImplicitTypeVariable:
 	""" Stand-in as the relevant type-expression for when the syntax doesn't bother. """
@@ -209,9 +214,13 @@ class Assumption(NamedTuple):
 	names: list[Nom]
 	type_expr: SimpleType
 
-def _bookend(head: Nom, coda: Nom):
-	if head.text != coda.text:
-		raise MismatchedBookendsError(head.head(), coda.head())
+def _bookend(head: Nom, where: Optional["WhereClause"]) -> Sequence["UserFunction"]:
+	if where is None: return ()
+	if head.text == where.coda.text:
+		return where.sub_fns
+	else:
+		raise MismatchedBookendsError(head.head(), where.coda.head())
+	
 
 class UserFunction(Term):
 	namespace: NS
@@ -231,15 +240,8 @@ class UserFunction(Term):
 		self.params = params or ()
 		self.result_type_expr = expr_type
 		self.expr = expr
-		if where:
-			_bookend(nom, where.coda)
-			self.where = where.sub_fns
-		else:
-			self.where = ()
+		self.where = _bookend(nom, where)
 			
-	def has_volatility(self):
-		return self.expr.is_volatile or any(f.has_volatility() for f in self.where)
-	
 	def head(self) -> slice:
 		return self.nom.head()
 	
@@ -265,28 +267,33 @@ class WhereClause(NamedTuple):
 	coda: Nom
 
 class UserAgent(Term):
-	field_space: NS
-	fields: Sequence[FormalParameter]
+	member_space: NS
+	members: Sequence[FormalParameter]
 	message_space: NS
 	def head(self) -> slice: return self.nom.head()
 	def __init__(
 		self,
 		nom: Nom,
-		fields: Optional[Sequence[FormalParameter]],
+		members: Optional[Sequence[FormalParameter]],
 		behaviors: Sequence["Behavior"],
 		coda: Nom,
 	):
 		super().__init__(nom)
-		self.fields = fields or ()
+		self.members = members or ()
 		self.behaviors = behaviors
-		_bookend(nom, coda)
+		if nom.text != coda.text:
+			raise MismatchedBookendsError(nom.head(), coda.head())
 
-	def field_names(self):
-		return [f.nom.text for f in self.fields]
+	def member_names(self):
+		return [f.nom.text for f in self.members]
 
 
 class Behavior(Term):
 	namespace: NS
+	params: Sequence[FormalParameter]
+	where: Sequence["UserFunction"]
+	reads_members: set[FormalParameter]  # Resolver must fill this.
+	
 	def head(self) -> slice: return self.nom.head()
 	def __repr__(self):
 		p = ", ".join(map(str, self.params))
@@ -296,15 +303,15 @@ class Behavior(Term):
 		nom: Nom,
 		params: Sequence[FormalParameter],
 		expr: ValExpr,
+		where: Optional["WhereClause"]
 	):
 		super().__init__(nom)
 		self.params = params or ()
 		for p in self.params: p.is_strict = True
 		self.expr = expr
+		self.where = _bookend(nom, where)
 
 class Literal(ValExpr):
-	is_volatile = False  # Not that this will ever be tested, but it's well to be consistent.
-	
 	def __init__(self, value: Any, a_slice: slice):
 		self.value, self._slice = value, a_slice
 	
@@ -321,21 +328,13 @@ class Lookup(ValExpr):
 	# Reminder: This AST node exists in opposition to TypeCall so I can write
 	# behavior for references in value context vs. references in type context.
 	ref:Reference
-	is_volatile = False
 	def __init__(self, ref: Reference): self.ref = ref
 	def head(self) -> slice: return self.ref.head()
 	def __str__(self): return str(self.ref)
 
-def is_self_reference(it):
-	if isinstance(it, Lookup):
-		if isinstance(it.ref, PlainReference):
-			return it.ref.nom.text == "SELF"
-	return False
-
 class FieldReference(ValExpr):
 	def __init__(self, lhs: ValExpr, field_name: Nom):
 		self.lhs, self.field_name = lhs, field_name
-		self.is_volatile = lhs.is_volatile or is_self_reference(lhs)
 	def __str__(self): return "(%s.%s)" % (self.lhs, self.field_name.text)
 	def head(self) -> slice: return self.field_name.head()
 
@@ -343,26 +342,22 @@ class BindMethod(ValExpr):
 	def __init__(self, receiver: ValExpr, _bang:Nom, method_name: Nom):
 		self.receiver, self.method_name = receiver, method_name
 		self._head = _bang.head()
-		self.is_volatile = receiver.is_volatile
 	def __str__(self): return "(%s.%s)" % (self.receiver, self.method_name.text)
 	def head(self) -> slice: return self._head
 
 class AsTask(ValExpr):
-	is_volatile = False
 	def __init__(self, head:slice, sub:ValExpr):
 		self._head = head
 		self.sub = sub
 	def head(self) -> slice: return self._head
 
 class Skip(ValExpr):
-	is_volatile = False
 	def __init__(self, head: slice): self._head = head
 	def head(self) -> slice: return self._head
 
 class Binary(ValExpr):
 	def __init__(self, lhs: ValExpr, op:Nom, rhs: ValExpr):
 		self.lhs, self.op, self.rhs = lhs, op, rhs
-		self.is_volatile = self.lhs.is_volatile or self.rhs.is_volatile
 	def head(self) -> slice: return self.op.head()
 
 class BinExp(Binary): pass
@@ -371,7 +366,6 @@ class ShortCutExp(Binary): pass
 class UnaryExp(ValExpr):
 	def __init__(self, op:Nom, arg: ValExpr):
 		self.op, self.arg = op, arg
-		self.is_volatile = arg.is_volatile
 
 	def head(self) -> slice: return self.op.head()
 
@@ -380,7 +374,6 @@ class Cond(ValExpr):
 	def __init__(self, then_part: ValExpr, _kw, if_part: ValExpr, else_part: ValExpr):
 		self._kw = _kw
 		self.then_part, self.if_part, self.else_part = then_part, if_part, else_part
-		self.is_volatile = any(x.is_volatile for x in (then_part, if_part, else_part))
 	def head(self) -> slice:
 		return self._kw
 
@@ -392,7 +385,6 @@ def CaseWhen(when_parts: list, else_part: ValExpr):
 class Call(ValExpr):
 	def __init__(self, fn_exp: ValExpr, args: list[ValExpr]):
 		self.fn_exp, self.args = fn_exp, args
-		self.is_volatile = self.fn_exp.is_volatile or any(a.is_volatile for a in args)
 	
 	def __str__(self):
 		return "%s(%s)" % (self.fn_exp, ', '.join(map(str, self.args)))
@@ -407,7 +399,6 @@ class ExplicitList(ValExpr):
 		for e in elts:
 			assert isinstance(e, ValExpr), e
 		self.elts = elts
-		self.is_volatile = any(e.is_volatile for e in elts)
 		
 	def head(self) -> slice:
 		return slice(self.elts[0].head().start, self.elts[-1].head().stop)
@@ -424,19 +415,11 @@ class Alternative:
 		self.pattern = pattern
 		self._head = _head
 		self.sub_expr = sub_expr
-		if where:
-			_bookend(pattern, where.coda)
-			self.where = where.sub_fns
-		else:
-			self.where = ()
+		self.where = _bookend(pattern, where)
 	def head(self) -> slice:
 		return self._head.head()
 	
-	def has_volatility(self):
-		return self.sub_expr.is_volatile or any(f.has_volatility() for f in self.where)
-
 class Absurdity(ValExpr):
-	is_volatile = False
 	def __init__(self, head:Nom, reason:Optional[Literal]):
 		self._head = head
 		self.reason = reason
@@ -484,7 +467,6 @@ class MatchExpr(ValExpr):
 		self.subject = subject
 		self.hint = hint
 		self.alternatives, self.otherwise = alternatives, otherwise
-		self.is_volatile = subject.expr.is_volatile or any(a.has_volatility() for a in alternatives) or (otherwise and otherwise.is_volatile)
 	
 	def head(self) -> slice:
 		return self.subject.head()
@@ -500,8 +482,6 @@ class DoBlock(ValExpr):
 	# The value of a do-block does not depend on when it runs.
 	# Its consequence may so depend, but by definition steps run in sequence.
 
-	is_volatile = False
-
 	def __init__(self, agents:list[NewAgent], _head:Nom, steps:list[ValExpr]):
 		self.agents = agents
 		self.steps = steps
@@ -510,7 +490,7 @@ class DoBlock(ValExpr):
 	def head(self) -> slice:
 		return self._head.head()
 
-class AssignField(Reference):
+class AssignMember(Reference):
 	def __init__(self, nom:Nom, expr:ValExpr):
 		super().__init__(nom)
 		self.expr = expr
@@ -525,7 +505,6 @@ class LambdaForm(ValExpr):
 		self._slice = slice(left.head().start, right.head().stop)
 		nom = Nom(_gensym(), self._slice)
 		self.function = UserFunction(nom, params, None, body, None)
-		self.is_volatile = self.function.has_volatility()
 		
 	def head(self):
 		return self._slice