+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Read this to get a general idea of what's new and nontrivial.
 
 ## April 2024
 
+* 4 Apr: Implemented snapshot semantics for actor member access. (Much code oddly simplified!)
+  Syntax now always uses ``my foo`` instead of ``self.foo`` to refer to an actor's own private state.
 * 3 Apr: I have decided on snapshot semantics for actor member access, with associated doc updates.
 * Various and sundry bugs fixed, mainly relating to recent features.